Rename: 크기를 %가 아닌 vw, vh로 변경하여, 파일 이름 수정
