--- src/components/public/atoms/FullSizeSection/FullSizeSection.tsx
@@ -1,14 +1,0 @@
-import { ReactNode } from "react";
-
-interface IFullSizeSection {
-  className?: string;
-  children: ReactNode;
-}
-
-const FullSizeSection = ({ className, children }: IFullSizeSection) => {
-  return (
-    <section className={`w-full h-full ${className ?? ""}`}>{children}</section>
-  );
-};
-
-export default FullSizeSection;
